style: Improve dashboard filter modal UX
